+++ example/lib/page/style/classical_page.dart
@@ -54,6 +54,7 @@ class _ClassicPageState extends State<ClassicPage> {
       body: EasyRefresh(
         clipBehavior: Clip.none,
         controller: _controller,
+        canLoadAfterNoMore: true,
         header: ClassicHeader(
           clamping: _headerProperties.clamping,
           backgroundColor: _headerProperties.background
@@ -113,7 +114,9 @@ class _ClassicPageState extends State<ClassicPage> {
                   return;
                 }
                 setState(() {
-                  _count += 5;
+                  if (_count < 20) {
+                    _count += 5;
+                  }
                 });
                 _controller.finishLoad(_count >= 20
                     ? IndicatorResult.noMore