+++ lib/src/easy_refresh.dart
@@ -106,10 +106,10 @@ class EasyRefresh extends StatefulWidget {
   final bool simultaneously;
 
   /// Is it possible to refresh after there is no more.
-  final bool noMoreRefresh;
+  final bool canRefreshAfterNoMore;
 
   /// Is it possible to load after there is no more.
-  final bool noMoreLoad;
+  final bool canLoadAfterNoMore;
 
   /// Reset after refresh when no more deactivation is loaded.
   final bool resetAfterRefresh;
@@ -183,8 +183,8 @@ class EasyRefresh extends StatefulWidget {
     this.notRefreshHeader,
     this.notLoadFooter,
     this.simultaneously = false,
-    this.noMoreRefresh = false,
-    this.noMoreLoad = false,
+    this.canRefreshAfterNoMore = false,
+    this.canLoadAfterNoMore = false,
     this.resetAfterRefresh = true,
     this.refreshOnStart = false,
     this.refreshOnStartHeader,
@@ -214,8 +214,8 @@ class EasyRefresh extends StatefulWidget {
     this.notRefreshHeader,
     this.notLoadFooter,
     this.simultaneously = false,
-    this.noMoreRefresh = false,
-    this.noMoreLoad = false,
+    this.canRefreshAfterNoMore = false,
+    this.canLoadAfterNoMore = false,
     this.resetAfterRefresh = true,
     this.refreshOnStart = false,
     this.refreshOnStartHeader,
@@ -346,13 +346,13 @@ class _EasyRefreshState extends State<EasyRefresh>
     // Update header and footer.
     _headerNotifier._update(
       indicator: _header,
-      noMoreProcess: widget.noMoreRefresh,
+      canProcessAfterNoMore: widget.canRefreshAfterNoMore,
       task: _onRefresh,
       waitTaskRefresh: _waitRefreshResult,
     );
     _footerNotifier._update(
       indicator: _footer,
-      noMoreProcess: widget.noMoreLoad,
+      canProcessAfterNoMore: widget.canLoadAfterNoMore,
       task: widget.onLoad,
       waitTaskRefresh: _waitLoadResult,
     );
@@ -381,7 +381,7 @@ class _EasyRefreshState extends State<EasyRefresh>
         userOffsetNotifier: userOffsetNotifier,
         vsync: this,
         onRefresh: _onRefresh,
-        noMoreRefresh: widget.noMoreRefresh,
+        canProcessAfterNoMore: widget.canRefreshAfterNoMore,
         waitRefreshResult: _waitRefreshResult,
         onCanRefresh: () {
           if (widget.simultaneously) {
@@ -396,7 +396,7 @@ class _EasyRefreshState extends State<EasyRefresh>
         userOffsetNotifier: userOffsetNotifier,
         vsync: this,
         onLoad: widget.onLoad,
-        noMoreLoad: widget.noMoreLoad,
+        canProcessAfterNoMore: widget.canLoadAfterNoMore,
         waitLoadResult: _waitLoadResult,
         onCanLoad: () {
           if (widget.simultaneously) {